+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPX8CSJ9RHEMT34SWEDP5NK3"
+        "01KPYYEZZXR138CPRBK2PW4MK7"
     )
     r"""The unique identifier of the data source"""