+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPX8CSJB2XGR7QCE5DDQHMEP"
+        "01KPYYF00010ZHX53PVHX15BMW"
     )
     r"""The unique identifier of the data source"""